fix: Fix middleware so the response_size_bytes metric is populated
Add wrapping of response body so we can count bytes.

This does buffer the requests to momory, which should be fine but could
be problematic for large requests. We should stream the data through but
I think it adds more complexity than it is worth, and that we should
only address it if it becomes an issue.